ARUHA-56: refactored configurations and tests for security;
