+++ src/main/java/de/zalando/aruha/nakadi/config/NakadiConfig.java
@@ -13,12 +13,8 @@ import com.ryantenney.metrics.spring.config.annotation.MetricsConfigurerAdapter;
 import de.zalando.aruha.nakadi.controller.EventPublishingController;
 import de.zalando.aruha.nakadi.controller.EventStreamController;
 import de.zalando.aruha.nakadi.controller.PartitionsController;
-import de.zalando.aruha.nakadi.repository.db.EventTypeDbRepository;
-import de.zalando.aruha.nakadi.repository.kafka.KafkaFactory;
-import de.zalando.aruha.nakadi.repository.kafka.KafkaLocationManager;
-import de.zalando.aruha.nakadi.repository.kafka.KafkaTopicRepository;
-import de.zalando.aruha.nakadi.repository.kafka.KafkaRepositorySettings;
-import de.zalando.aruha.nakadi.repository.zookeeper.ZooKeeperHolder;
+import de.zalando.aruha.nakadi.repository.EventTypeRepository;
+import de.zalando.aruha.nakadi.repository.TopicRepository;
 import org.json.JSONObject;
 import de.zalando.aruha.nakadi.service.EventStreamFactory;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -26,7 +22,6 @@ import org.springframework.boot.context.embedded.ServletRegistrationBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
-import org.springframework.core.env.Environment;
 import org.springframework.core.task.SimpleAsyncTaskExecutor;
 import org.springframework.core.task.TaskExecutor;
 import org.springframework.scheduling.annotation.EnableScheduling;
@@ -40,10 +35,10 @@ public class NakadiConfig {
     public static final MetricRegistry METRIC_REGISTRY = new MetricRegistry();
 
     @Autowired
-    private Environment environment;
-    
+    private EventTypeRepository eventTypeRepository;
+
     @Autowired
-    private EventTypeDbRepository eventTypeDbRepository;
+    private TopicRepository topicRepository;
 
     @Bean
     public TaskExecutor taskExecutor() {
@@ -83,38 +78,8 @@ public class NakadiConfig {
     }
 
     @Bean
-    public ZooKeeperHolder zooKeeperHolder() {
-        return new ZooKeeperHolder(
-                environment.getProperty("nakadi.zookeeper.brokers"),
-                environment.getProperty("nakadi.zookeeper.kafkaNamespace", ""),
-                environment.getProperty("nakadi.zookeeper.exhibitor.brokers"),
-                Integer.parseInt(environment.getProperty("nakadi.zookeeper.exhibitor.port", "0"))
-        );
-    }
-
-    @Bean
-    public KafkaLocationManager getKafkaLocationManager() {
-        return new KafkaLocationManager();
-    }
-
-    @Bean
-    public KafkaFactory kafkaFactory() {
-        return new KafkaFactory(getKafkaLocationManager());
-    }
-
-    @Bean
-    public KafkaRepositorySettings kafkaRepositorySettings() {
-        return new KafkaRepositorySettings();
-    }
-
-    @Bean
-    public KafkaTopicRepository kafkaRepository() {
-        return new KafkaTopicRepository(zooKeeperHolder(), kafkaFactory(), kafkaRepositorySettings());
-    }
-
-    @Bean
     public EventStreamController eventStreamController() {
-        return new EventStreamController(kafkaRepository(), jacksonObjectMapper(), eventStreamFactory());
+        return new EventStreamController(topicRepository, jacksonObjectMapper(), eventStreamFactory());
     }
 
     @Bean
@@ -124,12 +89,12 @@ public class NakadiConfig {
 
     @Bean
     public PartitionsController partitionsController() {
-        return new PartitionsController(kafkaRepository());
+        return new PartitionsController(topicRepository);
     }
 
     @Bean
     public EventPublishingController eventPublishingController() {
-        return new EventPublishingController(kafkaRepository(), eventTypeDbRepository);
+        return new EventPublishingController(topicRepository, eventTypeRepository);
     }
 
 }
