+++ src/main/java/de/zalando/aruha/nakadi/repository/db/EventTypeDbRepository.java
@@ -8,25 +8,21 @@ import de.zalando.aruha.nakadi.exceptions.NakadiException;
 import de.zalando.aruha.nakadi.repository.DuplicatedEventTypeNameException;
 import de.zalando.aruha.nakadi.repository.EventTypeRepository;
 import de.zalando.aruha.nakadi.exceptions.NoSuchEventTypeException;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DuplicateKeyException;
 import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.core.RowMapper;
-import org.springframework.stereotype.Component;
 
 import java.io.IOException;
 import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.util.List;
 
-@Component
 public class EventTypeDbRepository implements EventTypeRepository {
 
     private final JdbcTemplate jdbcTemplate;
     private final ObjectMapper jsonMapper;
 
-    @Autowired
     public EventTypeDbRepository(final JdbcTemplate jdbcTemplate, final ObjectMapper objectMapper) {
         this.jdbcTemplate = jdbcTemplate;
         this.jsonMapper = objectMapper;
