+++ src/main/resources/application.yml
@@ -43,3 +43,9 @@ nakadi:
     mode: BASIC
     tokenInfoUri: https://auth.zalando.com/oauth2/tokeninfo
     clientId: stups_aruha-event-store-poc
+    scopes:
+      uid: uid
+      nakadiRead: nakadi.read
+      eventTypeWrite: nakadi_event_type.write
+      eventStreamRead: nakadi_event_stream.read
+      eventStreamWrite: nakadi_event_stream.write
