+++ src/test/java/de/zalando/aruha/nakadi/security/AuthenticationTest.java
@@ -4,12 +4,17 @@ package de.zalando.aruha.nakadi.security;
 import com.google.common.collect.ArrayListMultimap;
 import com.google.common.collect.ImmutableSet;
 import com.google.common.collect.Multimap;
+import de.zalando.aruha.nakadi.Application;
 import de.zalando.aruha.nakadi.config.SecuritySettings;
+import de.zalando.aruha.nakadi.repository.EventTypeRepository;
 import de.zalando.aruha.nakadi.repository.TopicRepository;
 import de.zalando.aruha.nakadi.repository.db.EventTypeDbRepository;
 import org.junit.Before;
+import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Value;
+import org.springframework.boot.test.SpringApplicationConfiguration;
+import org.springframework.boot.test.WebIntegrationTest;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
@@ -18,6 +23,9 @@ import org.springframework.security.core.authority.AuthorityUtils;
 import org.springframework.security.oauth2.provider.OAuth2Authentication;
 import org.springframework.security.oauth2.provider.OAuth2Request;
 import org.springframework.security.oauth2.provider.token.ResourceServerTokenServices;
+import org.springframework.test.annotation.DirtiesContext;
+import org.springframework.test.context.ActiveProfiles;
+import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultMatcher;
 import org.springframework.test.web.servlet.request.MockHttpServletRequestBuilder;
@@ -36,8 +44,14 @@ import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
 import static org.springframework.http.HttpStatus.FORBIDDEN;
 import static org.springframework.http.HttpStatus.UNAUTHORIZED;
+import static org.springframework.test.annotation.DirtiesContext.ClassMode.AFTER_CLASS;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
+@RunWith(SpringJUnit4ClassRunner.class)
+@SpringApplicationConfiguration(Application.class)
+@WebIntegrationTest
+@DirtiesContext(classMode = AFTER_CLASS)
+@ActiveProfiles("test")
 public abstract class AuthenticationTest {
 
     @Configuration
@@ -71,7 +85,6 @@ public abstract class AuthenticationTest {
         }
 
         @Bean
-        @Primary
         public ResourceServerTokenServices mockResourceTokenServices() {
             final ResourceServerTokenServices tokenServices = mock(ResourceServerTokenServices.class);
 
@@ -90,24 +103,19 @@ public abstract class AuthenticationTest {
         }
 
         @Bean
-        @Primary
         public SecuritySettings mockSecuritySettings() {
             final SecuritySettings settings = mock(SecuritySettings.class);
             when(settings.getAuthMode()).thenReturn(authMode);
-            when(settings.getClientId()).thenReturn("foo");
-            when(settings.getTokenInfoUri()).thenReturn("http://foo.bar");
             return settings;
         }
 
         @Bean
-        @Primary
         public TopicRepository mockTopicRepository() {
             return mock(TopicRepository.class);
         }
 
         @Bean
-        @Primary
-        public EventTypeDbRepository mockDbRepository() {
+        public EventTypeRepository mockDbRepository() {
             return mock(EventTypeDbRepository.class);
         }
     }
