+++ src/test/java/de/zalando/aruha/nakadi/security/OffModeAuthenticationTest.java
@@ -2,26 +2,15 @@ package de.zalando.aruha.nakadi.security;
 
 
 import com.google.common.collect.ImmutableList;
-import de.zalando.aruha.nakadi.Application;
 import de.zalando.aruha.nakadi.config.SecuritySettings;
 import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.SpringApplicationConfiguration;
-import org.springframework.boot.test.WebIntegrationTest;
-import org.springframework.test.annotation.DirtiesContext;
-import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 
 import java.util.List;
 
 import static org.springframework.http.HttpMethod.GET;
 import static org.springframework.http.HttpMethod.POST;
 import static org.springframework.http.HttpMethod.PUT;
-import static org.springframework.test.annotation.DirtiesContext.ClassMode.AFTER_CLASS;
 
-@RunWith(SpringJUnit4ClassRunner.class)
-@SpringApplicationConfiguration(Application.class)
-@WebIntegrationTest
-@DirtiesContext(classMode = AFTER_CLASS)
 public class OffModeAuthenticationTest extends AuthenticationTest {
 
     static {
